Fix error when user doesn't have any roles

+++ client/src/utils/modules/InviteBlockerModule.ts
@@ -14,7 +14,7 @@ export default class InviteBlockerModule {
             const guildInfo = await this.bot.database.getGuild(guildId);
             const context = new ChatInputMessageContext(message);
             const authorRoles = context.authorRoles.map(role => role.toString().replace("n", ""));
-            
+
             if (message.authorId === this.bot.applicationId || message.isFromBot) return;
             if (!inviteRegex.test(message.content)) return;
             if (authorRoles.find(role => guildInfo.InviteBlockerModule.whitelistedRoles.includes(role))) return;
@@ -28,7 +28,7 @@ export default class InviteBlockerModule {
             if (blockMessage.includes("{channel}")) {
                 blockMessage = blockMessage.replace("{channel}", `<#${message.channelId}>`);
             }
-            if (inviteRegex.test(message.content)) {
+            if (inviteRegex.test(message.content) && !authorRoles.find(role => guildInfo.InviteBlockerModule.whitelistedRoles.includes(role))) {
                 context.DeleteMessage(message.id, "Invite Blocker");
                 context.FoxyReply({
                     content: blockMessage
